Move properties to a separate unicode file
